Refactor k-link
