+++ panel/src/components/Navigation/ButtonNative.vue
@@ -23,7 +23,7 @@
 </template>
 
 <script>
-import tab from "@/mixins/tab.js";
+import tab from "@/ui/mixins/tab.js";
 
 export default {
   mixins: [tab],
