+++ panel/src/ui/mixins/tab.js
